Modularize SofaMiscEngine
