+++ modules/SofaMiscEngine/src/SofaMiscEngine/config.h.in
@@ -19,22 +19,13 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#ifndef SOFA_COMPONENT_MISC_ENGINE_INIT_H
-#define SOFA_COMPONENT_MISC_ENGINE_INIT_H
-#include "config.h"
+#pragma once
 
-namespace sofa
-{
+#include <sofa/config/sharedlibrary_defines.h>
 
-namespace component
-{
-
-
-void SOFA_MISC_ENGINE_API initMiscEngine();
-
-} // namespace component
-
-} // namespace sofa
-
+#ifdef SOFA_BUILD_MISC_ENGINE
+#cmakedefine SOFA_TARGET @PROJECT_NAME@
+#  define SOFA_SOFAMISCENGINE_API SOFA_EXPORT_DYNAMIC_LIBRARY
+#else
+#  define SOFA_SOFAMISCENGINE_API SOFA_IMPORT_DYNAMIC_LIBRARY
 #endif
-
