+++ modules/SofaMiscEngine/src/SofaMiscEngine/initMiscEngine.cpp
@@ -19,65 +19,21 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-//#define SOFA_COMPONENT_ENGINE_DisplacementMatrixEngine_CPP
+#include <sofa/helper/system/config.h>
+#include <SofaMiscEngine/initMiscEngine.h>
 
-#include "DisplacementMatrixEngine.inl"
 
-#include <sofa/core/ObjectFactory.h>
-
-namespace sofa
+namespace sofa::component
 {
 
-namespace component
+
+void initMiscEngine()
 {
-
-namespace engine
-{
-
-using namespace defaulttype;
-
-int DisplacementTransformEngineClass = core::RegisterObject("Converts a vector of Rigid to a vector of displacement transforms.")
-    .add< DisplacementTransformEngine<Rigid3Types,Mat4x4f> >()
-    .add< DisplacementTransformEngine<Rigid3Types,Rigid3Types::Coord> >()
-;
-
-template class SOFA_MISC_ENGINE_API DisplacementTransformEngine<Rigid3Types,Mat4x4f>;
-template class SOFA_MISC_ENGINE_API DisplacementTransformEngine<Rigid3Types,Rigid3Types::Coord>;
-
-template <>
-void DisplacementTransformEngine<Rigid3Types,Rigid3Types::Coord>::setInverse( Rigid3Types::Coord& inv, const Coord& x0 )
-{
-    inv = Rigid3Types::inverse(x0);
+    static bool first = true;
+    if (first)
+    {
+        first = false;
+    }
 }
 
-template <>
-void DisplacementTransformEngine<Rigid3Types,Rigid3Types::Coord>::mult( Rigid3Types::Coord& out, const Rigid3Types::Coord& inv, const Coord& x )
-{
-    out = x;
-    out.multRight(inv);
-}
-
-template <>
-void DisplacementTransformEngine<Rigid3Types,Mat4x4f>::setInverse( Mat4x4f& inv, const Coord& x0 )
-{
-    Rigid3Types::inverse(x0).toMatrix(inv);
-}
-
-template <>
-void DisplacementTransformEngine<Rigid3Types,Mat4x4f>::mult( Mat4x4f& out, const Mat4x4f& inv, const Coord& x )
-{
-    x.toMatrix(out);
-    out = out * inv;
-}
-
-/////////////////////////////////////////
-
-int DisplacementMatrixEngineClass = core::RegisterObject("Converts a vector of Rigid to a vector of displacement matrices.")
-    .add< DisplacementMatrixEngine<Rigid3Types> >()
-;
-
-} // namespace engine
-
-} // namespace component
-
-} // namespace sofa
+} // namespace sofa::component
